+++ turbo/packages/core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [8.603.2](https://github.com/vm0-ai/vm0/compare/core-v8.603.1...core-v8.603.2) (2026-08-30)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @okouai/api-contracts bumped to 1.381.10
+
 ## [8.603.1](https://github.com/vm0-ai/vm0/compare/core-v8.603.0...core-v8.603.1) (2026-08-30)